# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a playground Web API using Gin framework.",
        "title": "Go Web API Playground",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/examples/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lyrics/{artist}/{song}": {
            "get": {
                "description": "Get lyrics from the public api https://api.lyrics.ovh/v1/",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lyrics"
                ],
                "summary": "Get lyrics from an Artist and a Song",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist",
                        "name": "artist",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Song",
                        "name": "song",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SongLyrics"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/startrek/crewmember": {
            "get": {
                "description": "Get all Crew Members",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "startrek"
                ],
                "summary": "Get all Crew Members",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Crew Member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "startrek"
                ],
                "summary": "Create a Crew Member",
                "parameters": [
                    {
                        "description": "Crew member",
                        "name": "crewMember",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CrewMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/startrek/crewmember/{id}": {
            "get": {
                "description": "Get a Crew Member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "startrek"
                ],
                "summary": "Get a Crew Member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CrewMember"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Crew Member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "startrek"
                ],
                "summary": "Update a Crew Member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Crew member",
                        "name": "crewMember",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CrewMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CrewMember"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Crew Member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "startrek"
                ],
                "summary": "Delete a Crew Member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CrewMember": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "starship": {
                    "type": "string"
                }
            }
        },
        "models.SongLyrics": {
            "type": "object",
            "properties": {
                "Artist": {
                    "type": "string"
                },
                "Lyrics": {
                    "type": "string"
                },
                "SongTitle": {
                    "type": "string"
                }
            }
        }
    }
}